upload une ou plusieurs images + affichage dans toutes les cartes et barre de recherche

+++ backend/routes/deleteproduit.js
@@ -1,25 +1,28 @@
 import express from "express";
 import Produits from "../models/Produits.js";
+import jwt from "jsonwebtoken";
+import Images from "../models/Images.js";
 
 const router = express.Router();
 router.use(express.json());
-//import jwt from "jsonwebtoken";
 
 router.delete("/", async (req, res) => {
   try {
-    //const{codeProduit} = req.params;
-    const { codeProduit } = req.body;
-    //const user = jwt.verify(token,process.env.JWT_SECRET);
-    /*if(user.role !== "admin"){
-            res.status(401).json({
-                message: "Vous n'avez pas les autorisations nécessaires pour ajouter un produit."
-            });
-            return;
-        }*/
+    const { codeProduit, token } = req.body;
+
+    const user = jwt.verify(token, process.env.JWT_SECRET);
+    if (user.role !== "admin") {
+      res.status(401).json({
+        message:
+          "Vous n'avez pas les autorisations nécessaires pour ajouter un produit.",
+      });
+      return;
+    }
 
     const produitSupp = await Produits.findOneAndDelete({ codeProduit });
 
     if (produitSupp) {
+      await Images.find({ codeProduit }).deleteMany();
       res
         .status(200)
         .json({ message: "Le produit a été supprimé avec succès !" });
